Updated folder structure, moved test code around, updated imports to reflect new structure
