+++ SpriteTest/spaceship_sprite.png
