+++ sprite/imagesheet.py
